requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-ckeditor==5.7.1
django-cors-headers==3.1.1
django-js-asset==1.2.2
django-localflavor==2.2
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
openapi-codec==1.3.2
packaging==19.2
pyparsing==2.4.2
python-stdnum==1.12
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
